requirements.txt
opencv-python
pillow
pip-chill
pynmea2
pyproj
pyserial
rasterio
